% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Gaussian_Fold.R
\name{a}
\alias{a}
\title{Calculating the mixture pdf of a Fold Gaussian kernel functions}
\usage{
a(x)
}
\arguments{
\item{x}{A numerical value.}

\item{data}{A value or a vector.}

\item{h}{A numerical value for bandwidth .}
}
\value{
A numerical value.
}
\description{
Calculating the mixture pdf of a Fold Gaussian kernel functions
}
\examples{
gaussian.fold.pdf(x, data, h);
}
