% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Gaussian.R
\name{cdf.gaussian}
\alias{cdf.gaussian}
\title{Cdf for Gaussian kernel functions}
\usage{
cdf.gaussian(x)
}
\arguments{
\item{x}{A numerical value.}
}
\value{
A numerical.
}
\description{
Cdf for Gaussian kernel functions
}
\examples{
cdf.gaussian(10);
}
