% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_kernel.R
\name{cdf.kernel}
\alias{cdf.kernel}
\title{Cumulative distribution function for any kernel function}
\usage{
cdf.kernel(x, data, bandwidth = 1, kernel_name = "gaussian")
}
\arguments{
\item{x}{a vector/ numeric the points in which we want to estimate the probability}

\item{data}{a vector the windows}

\item{bandwidth}{a vector/ numeric of bandwidth. If one bandwidth is given to estimate the pdf of vector
the bandwidth will be duplicated. If the length of vector of bandiwdth is not equal to
the length of x, error will come out}

\item{kernel_name}{a string}
}
\value{
cdf: a vector/ numerical values of the estimated PDF

x: the values of points

bandwidth: bandwidth
}
\description{
Cumulative distribution function for any kernel function
}
\examples{
x = c(0.11, 0.3);
data = c(0.2, 0.3);
bandwidth = 1;
cdf.kernel(x, data, bandwidth, kernel_name = "gaussian");
}
