% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loss_functions.R
\name{crps}
\alias{crps}
\title{Continuous rank probability score (crps) for any kernel distribution}
\usage{
crps(x, data, bandwidth, kernel_name)
}
\arguments{
\item{x}{the data to evaluate at (numeric/ vector)}

\item{data}{the data (numeric/ vector)}

\item{bandwidth}{A numerical value}

\item{kernel_name}{A string}
}
\value{
x a vector of x

kernel a string

bandwidth a numerical value

crps A vector of CRPS evaluate at x

Average CRPS The mean of the CRPS
}
\description{
Compute the negative continuous rank probabilistic score for Gaussian kernel for a vector using a single bandwidth
}
\examples{
data <- seq(0.1, 0.5, by = 0.1);
x <- c(0.1, 0.2, 0.3,  0.15);
bandwidth <- 0.1;
gaussian.crps (x, data, bandwidth);
}
