% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Epanechnikov.R
\name{epanechnikov.cdf}
\alias{epanechnikov.cdf}
\title{Calculating cumulative probability functions for Epanechnikov kernel}
\usage{
epanechnikov.cdf(x)
}
\arguments{
\item{x}{vector of observations.}
}
\value{
A vector of score values.
examples:
epanechnikov.cdf(0.45);
}
\description{
Calculating cumulative probability functions for Epanechnikov kernel
}
