% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Epanechnikov.R
\name{epanechnikov.kernel}
\alias{epanechnikov.kernel}
\title{Calculating probability density functions for Epanechnikov kernel}
\usage{
epanechnikov.kernel(x)

epanechnikov.kernel(x)
}
\arguments{
\item{x}{vector of observations.}
}
\value{
A vector of score values.
examples:
epanechnikov.kernel(0.45);

A vector of score values.
examples:
epanechnikov.p2norm(0.45);
}
\description{
Calculating probability density functions for Epanechnikov kernel

Calculating L2-norm of for Epanechnikov kernel
}
