% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Epanechnikov.R
\name{epanechnikov.p2norm}
\alias{epanechnikov.p2norm}
\title{L2-norm Epanechnikov kernel}
\usage{
epanechnikov.p2norm(x)
}
\arguments{
\item{x}{vector of observations.}
}
\value{
A vector of score values.
examples:
epanechnikov.p2norm(0.45);
}
\description{
Calculating L2-norm of Epanechnikov kernel from negative infinity to infinity
}
