% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Gaussian.R
\name{gaussian.ccdf}
\alias{gaussian.ccdf}
\title{Cumulative cdf of Gaussian Kernel Function}
\usage{
gaussian.ccdf(x)
}
\arguments{
\item{x}{A numerical value.}
}
\value{
A numerical.
}
\description{
This function calculates the integral of a Gaussian kernel cdf from
negative infinity up until a value
}
\examples{
gaussian.ccdf(10);
}
