% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Gaussian.R
\name{gaussian.cdf}
\alias{gaussian.cdf}
\title{Cdf for Gaussian kernel functions}
\usage{
gaussian.cdf(x)
}
\arguments{
\item{x}{A numerical value.}
}
\value{
A numerical.
}
\description{
Cdf for Gaussian kernel functions
}
\examples{
gaussian.cdf(10);
}
