% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Gaussian_Fold.R
\name{gaussian.fold.cdf}
\alias{gaussian.fold.cdf}
\title{Calculate the cdf of a folded Gaussian kernel from 0 to infinity}
\usage{
gaussian.fold.cdf(x)
}
\arguments{
\item{x}{The value of x}
}
\value{
a numerical / vector
}
\description{
Calculate the cdf of a folded Gaussian kernel from 0 to infinity
}
\examples{
gaussian.fold.cdf(10);
}
