% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Gaussian_Fold.R
\name{gaussian.fold.kernel}
\alias{gaussian.fold.kernel}
\title{Calculating the pdf of a fold Gaussian kernel functions}
\usage{
gaussian.fold.kernel(x)
}
\arguments{
\item{x}{A numerical value.}
}
\value{
A numerical.
}
\description{
Calculating the pdf of a fold Gaussian kernel functions
}
\examples{
gaussian.fold.kernel(10);
}
