% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Gaussian_Fold.R
\name{gaussian.fold.mean}
\alias{gaussian.fold.mean}
\title{Calculate the mean (expected) value, E|X|,  of a folded Gaussian kernel from 0 to infinity}
\usage{
gaussian.fold.mean(x)
}
\arguments{
\item{x}{The value of x}
}
\value{
a numerical / vector
}
\description{
Calculate the mean (expected) value, E|X|,  of a folded Gaussian kernel from 0 to infinity
}
\examples{
gaussian.fold.mean(10);
}
