% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Gaussian.R
\name{gaussian.kernel}
\alias{gaussian.kernel}
\title{Pdf of Gaussian kernel functions}
\usage{
gaussian.kernel(x)
}
\arguments{
\item{x}{A numerical value.}
}
\value{
A numerical.
}
\description{
Pdf of Gaussian kernel functions
}
\examples{
gaussian.kernel(10);
}
