% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Gaussian.R
\name{gaussian.p2norm}
\alias{gaussian.p2norm}
\title{L2-norm of a Gaussian kernel}
\usage{
gaussian.p2norm(data)
}
\arguments{
\item{x}{A numerical value.}
}
\value{
A numerical.
}
\description{
This function calculates the p2-norm of Gaussian kernel
}
\examples{
gaussian.p2norm(10);
}
