% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradient_descent.R
\name{gradient_descent}
\alias{gradient_descent}
\title{Gradient descent for Gaussian logloss}
\usage{
gradient_descent(
  x,
  data,
  bandwidth,
  alpha = 0.01,
  max_iter = 1000,
  tolerance = 1e-06
)
}
\arguments{
\item{x}{the data to evaluate at (numeric/ vector)}

\item{data}{the data (numeric/ vector)}

\item{bandwidth}{A numerical value}
}
\value{
value
}
\description{
Compute the derivative of negative logarithmic score
            for Gaussian kernel for a vector using a single bandwidth
}
\examples{
data <- seq(0.1, 1, by = 0.1)
x <- c(0.1, 0.2, 0.3,  0.15)
bandwidth = 0.01
gradient_descent(x, data, bandwidth)
}
