% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradient_logloss.R
\name{gradient_logloss}
\alias{gradient_logloss}
\title{Gradient logloss}
\usage{
gradient_logloss(x, data, bandwidth)
}
\arguments{
\item{x}{the data to evaluate at (numeric/ vector)}

\item{data}{the data (numeric/ vector)}

\item{bandwidth}{A numerical value}
}
\value{
value
}
\description{
Compute the derivative of negative logarithmic score
            for Gaussian kernel for a vector using a single bandwidth
}
\examples{
data <- seq(0.1, 0.5, by = 0.1)
x <- c(0.1, 0.2, 0.3,  0.15)
bandwidth = 0.1
gradient_logloss(x, data, bandwidth)
}
