% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Gaussian.R
\name{kernel.gaussian}
\alias{kernel.gaussian}
\title{Pdf of Gaussian kernel functions}
\usage{
kernel.gaussian(x)
}
\arguments{
\item{x}{A numerical value.}
}
\value{
A numerical.
}
\description{
Pdf of Gaussian kernel functions
}
\examples{
kernel.gaussian(10);
}
