% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Gaussian.R
\name{kernel_pdf.gaussian}
\alias{kernel_pdf.gaussian}
\title{Calculating the pdf of the Gaussian kernel functions}
\usage{
\method{kernel_pdf}{gaussian}(obj, x, data)
}
\arguments{
\item{x}{A numerical value}

\item{data}{A vector of values}

\item{h}{bandwidth}
}
\value{
A list
}
\description{
Calculating the pdf of the Gaussian kernel functions
}
\examples{
x <- 1
data <- rnorm(10, 2, 1)
h <- 0.5
kernel_pdf.gaussian(obj, x, data)
}
