% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logloss.R
\name{log_loss}
\alias{log_loss}
\title{Log-loss}
\usage{
log_loss(x, data, bandwidth, kernel_name)
}
\arguments{
\item{x}{the data to evaluate at (numeric/ vector)}

\item{data}{the data (numeric/ vector)}

\item{bandwidth}{A numerical value}

\item{kernel_name}{A string}
}
\value{
x a vector of x

kernel_name

bandwidth

logloss A vector of loss evaluate at x

Average_loss The mean of the logloss
}
\description{
Compute the negative logarithmic score
            for Gaussian kernel for a vector using a single bandwidth
}
\examples{
data <- seq(0.1, 0.5, by = 0.1)
x <- c(0.1, 0.2, 0.3,  0.15)
bandwidth = 0.1
kernel_name = "gaussian"
log_loss(x=x, data, bandwidth, kernel_name= kernel_name)
}
