% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Logistic.R
\name{Logistic.cdf.p2Norm}
\alias{Logistic.cdf.p2Norm}
\title{Calculate the p2Norm of a Logistic kernel cdf from negative infinity to a value a}
\usage{
Logistic.cdf.p2Norm(x, a)
}
\arguments{
\item{x}{The value of x}

\item{a}{A numeric for the upper found of the integral}
}
\value{
a numerical / vector
}
\description{
Calculate the p2Norm of a Logistic kernel cdf from negative infinity to a value a
}
