% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Logistic.R
\name{Logistic.p2Norm}
\alias{Logistic.p2Norm}
\title{Calculate the p2Norm of a Logistic kernel from negative infinity to infinity}
\usage{
Logistic.p2Norm(x)
}
\arguments{
\item{x}{The value of x}
}
\value{
a numerical / vector
}
\description{
Calculate the p2Norm of a Logistic kernel from negative infinity to infinity
}
