% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Gaussian.R
\name{logloss.gaussian}
\alias{logloss.gaussian}
\title{Logloss for  Gaussian Kernel Function}
\usage{
logloss.gaussian(x)
}
\arguments{
\item{x}{A numerical value.}
}
\value{
A numerical.
}
\description{
This function calculates the log-loss of a Gaussian kernel cdf from
negative infinity up until a value
}
\examples{
logloss.gaussian(10);
}
