% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_Gaussian.R
\name{p2norm.gaussian}
\alias{p2norm.gaussian}
\title{L2-norm of a Gaussian kernel}
\usage{
p2norm.gaussian(data)
}
\arguments{
\item{x}{A numerical value.}
}
\value{
A numerical.
}
\description{
This function calculates the p2-norm of Gaussian kernel
}
\examples{
p2norm.gaussian(10);
}
