% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2norm_pdf.R
\name{p2norm_pdf}
\alias{p2norm_pdf}
\title{L2 - norm of kernel probability density function}
\usage{
p2norm_pdf(data, bandwidth = 1, kernel_name = "gaussian")
}
\arguments{
\item{data}{a vector the windows}

\item{bandwidth}{a numeric value of bandwidth. The default value is 1.}

\item{kernel_name}{a string. The default value is "gaussian"}
}
\value{
data: Vector of data input

bandwidth: bandwidth

kernel_name: name of the kernel used

p2norm: a vector/ numerical values of the estimated PDF
}
\description{
This function computes the L2-norm of the kernel probability density function.
\deqn{
L_{2} = \int_{-\inf}^{\inf} (f(x))^2 \ dx
}
where \eqn{f(x) = \frac{1}{nh}\sum_{i=1}^{n} K\left(\frac{x - X_{i}}{h}\right)}
}
\examples{
x = c(0.11)
data = c(0.2, 0.3)
bandwidth = 1
kernel_name = "gaussian"
p2norm_pdf(data = data , bandwidth = bandwidth, kernel_name)
}
