% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_kernel_variable_bandwidth.R
\name{pdf.kde}
\alias{pdf.kde}
\title{Probability density kernel function (pdf)}
\usage{
\method{pdf}{kde}(x, data, bandwidth, kernel_name)
}
\arguments{
\item{x}{a vector/ numeric the points in which we want to estimate the probability}

\item{data}{a vector the windows}

\item{bandwidth}{a vector/ numeric of bandwidth. If one bandwidth is given to estimate the pdf of vector
the bandwidth will be duplicated to the number of vector in. If the length of vector of bandwidth is not equal to
the length of x, error will come out}

\item{kernel_name}{a string}
}
\value{
pdf: a vector/ numerical values of the estimated PDF

x: the values of points

bandwidth: a numerical / vector value bandwidth
}
\description{
Probability density kernel function (pdf)
}
\examples{
x = c(0.11, 0.3);
data = c(0.2, 0.3);
bandwidth = c(0.5, 0.4);
pdf.kde(x, data, bandwidth, kernel_name = "gaussian");
}
