% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_kernel.R
\name{pdf_kernel}
\alias{pdf_kernel}
\title{Probability density function (pdf) for any kernel function}
\usage{
pdf_kernel(x, data, bandwidth = 1, kernel_name = "gaussian")
}
\arguments{
\item{x}{a vector/ numeric the points in which we want to estimate the probability}

\item{data}{a vector the windows}

\item{bandwidth}{a vector/ numeric of bandwidth. If one bandwidth is given to estimate the pdf of vector
the bandwidth will be duplicated. Default value is 1}

\item{kernel_name}{a string, default is "gaussian"}
}
\value{
pdf: a vector/ numerical values of the estimated PDF

x: the values of points

bandwidth: a numerical value
}
\description{
This function computes the probability density function (pdf) for kernel
}
\examples{
x = c(0.11);
data = c(0.2, 0.3);
bandwidth = 0.1;
pdf_kernel(x, data, bandwidth, kernel_name = "gaussian");
}
